udpate: participant apis

+++ backend/src/chat/services/chatroom/chatroom.service.ts
@@ -335,6 +335,16 @@ export class ChatroomService implements IChatroomService {
     const chatroom = await this.findChatroomByIdOrFail(chatroomId);
     const user = await this.findUserByIdOrFail(userId);
     const member = await this.findMemberById(userId, chatroomId);
+    // if (member && member.bannedAt !== null) {
+    //   console.log(
+    //     `User is banned from the chatroom of id${chatroomId}`,
+    //     member,
+    //   );
+    //   throw new UnauthorizedException(
+    //     `User is banned from the chatroom of id:${chatroomId}`,
+    //   );
+    // } else
+
     if (member) {
       console.log(
         `User already exists in the chatroom of id:${chatroomId}`,
@@ -351,12 +361,12 @@ export class ChatroomService implements IChatroomService {
       Chatroom: chatroom,
       User: user,
     });
-    // console.log('newMember:', newMember);
+    console.log('newMember:', newMember);
     const savedMember = await this.chatMemebrRepository.save(newMember);
-    // console.log('savedMember:', savedMember);
+    console.log('savedMember:', savedMember);
     this.chatEventsGateway.server.emit('newMemberList', savedMember);
     // return chatroomMember;
-    // this.postParticipants(userId, chatroomId);
+    this.postParticipants(userId, chatroomId);
   }
 
   async updateMemberInfo(
@@ -439,11 +449,11 @@ export class ChatroomService implements IChatroomService {
   async deleteMembers(userId: number, chatroomId: number) {
     await this.findChatroomByIdOrFail(chatroomId);
     const member = await this.findMemberByIdOrFail(userId, chatroomId);
-    // if (member.mutedAt === null) {
-    const removedMember = await this.chatMemebrRepository.remove(member);
-    console.log('removed member:', removedMember);
-    this.chatEventsGateway.server.emit('newMemberList', removedMember);
-    // }
+    if (member.mutedAt === null) {
+      const removedMember = await this.chatMemebrRepository.remove(member);
+      console.log('removed member:', removedMember);
+      this.chatEventsGateway.server.emit('newMemberList', removedMember);
+    }
     this.chatEventsGateway.server.emit('newMemberList');
   }
 
@@ -470,13 +480,13 @@ export class ChatroomService implements IChatroomService {
     const participant = await this.findParticipantById(userId, chatroomId);
     if (participant) {
       console.log(
-        `User already participate in the chatroom of id:${chatroomId}`,
+        `User already exists in the chatroom of id:${chatroomId}`,
         participant,
       );
-      // return;
-      throw new BadRequestException(
-        `User already participate in the chatroom of id:${chatroomId}`,
-      );
+      return;
+      // throw new BadRequestException(
+      //   `User already exists in the chatroom of id:${chatroomId}`,
+      // );
     }
 
     const newParticipant = this.chatParticipantRepository.create({
@@ -489,7 +499,6 @@ export class ChatroomService implements IChatroomService {
     const savedParticipant = await this.chatParticipantRepository.save(
       newParticipant,
     );
-    // this.chatEventsGateway.server.emit('newMemberList', savedParticipant);
     // console.log('saved Participant:', savedParticipant);
     // this.chatEventsGateway.server.emit('newMemberList', savedMember);
     return savedParticipant;
@@ -501,13 +510,13 @@ export class ChatroomService implements IChatroomService {
       userId,
       chatroomId,
     );
-    // if (participant.mutedAt === null) {
-    const removedParticipant = await this.chatParticipantRepository.remove(
-      participant,
-    );
-    console.log('removed participant:', removedParticipant);
-    // this.chatEventsGateway.server.emit('new participantList', removedparticipant);
-    // }
+    if (participant.mutedAt === null) {
+      const removedParticipant = await this.chatParticipantRepository.remove(
+        participant,
+      );
+      console.log('removed participant:', removedParticipant);
+      // this.chatEventsGateway.server.emit('new participantList', removedparticipant);
+    }
     // this.chatEventsGateway.server.emit('new participantList');
   }
 
@@ -528,10 +537,10 @@ export class ChatroomService implements IChatroomService {
     let updatedParticipant = null;
     // owner verification
     if (participant.userId !== chatroom.ownerId) {
-      throw new UnauthorizedException(
-        `User of id:${userId} is not an admin of chatroom of id:${chatroomId}`,
-      );
-      // return;
+      // throw new UnauthorizedException(
+      //   `User of id:${userId} is not an admin of chatroom of id:${chatroomId}`,
+      // );
+      return;
     }
     // mute
     if (updateMemberDto.mute === true) {
@@ -554,11 +563,19 @@ export class ChatroomService implements IChatroomService {
       updatedParticipant = await this.chatParticipantRepository.save(
         targetUser,
       );
-      this.chatEventsGateway.server.emit('mute', {
-        chatroomId: chatroomId,
-        targetUserId: targetUser.userId,
-      });
     }
+    // if (updateMemberDto.mute === true) {
+    //   this.addNewTimeout(
+    //     `${targetUser.id}_muted`,
+    //     targetUser.userId,
+    //     chatroomId,
+    //     15000,
+    //   );
+    //   targetUser.mutedAt = new Date();
+    //   updatedParticipant = await this.chatParticipantRepository.save(
+    //     targetUser,
+    //   );
+    // }
     console.log('updated participant:', updatedParticipant);
     return updatedParticipant;
   }
